+++ spec/PimEnterprise/Bundle/DataGridBundle/Datagrid/Proposal/GridHelperSpec.php
@@ -5,13 +5,14 @@ namespace spec\PimEnterprise\Bundle\DataGridBundle\Datagrid\Proposal;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Oro\Bundle\UserBundle\Entity\UserManager;
 use PhpSpec\ObjectBehavior;
+use Symfony\Component\Security\Core\SecurityContextInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 class GridHelperSpec extends ObjectBehavior
 {
-    function let(UserManager $userManager, ObjectRepository $userRepository)
+    function let(UserManager $userManager, SecurityContextInterface $securityContext, ObjectRepository $userRepository)
     {
-        $this->beConstructedWith($userManager);
+        $this->beConstructedWith($userManager, $securityContext);
 
         $userManager->getRepository()->willReturn($userRepository);
     }