+++ src/PimEnterprise/Bundle/WorkflowBundle/DependencyInjection/PimEnterpriseWorkflowExtension.php
@@ -32,6 +32,10 @@ class PimEnterpriseWorkflowExtension extends Extension
         $loader->load('appliers.yml');
         $loader->load('builder.yml');
         $loader->load('comparators.yml');
+        $loader->load('connector/array_converters.yml');
+        $loader->load('connector/processors.yml');
+        $loader->load('connector/readers.yml');
+        $loader->load('connector/writers.yml');
         $loader->load('controllers.yml');
         $loader->load('datagrid_listeners.yml');
         $loader->load('entities.yml');